requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Essential server bindings
gunicorn>=21.2.0
Flask>=3.0.0
Werkzeug>=3.0.0

# Headless Vision (Crucial for Linux without GUI)
opencv-python-headless>=4.8.0

# Deep Learning (CPU Only to save 1GB space)
torch>=2.0.0
torchvision>=0.15.0

# YOLO
ultralytics>=8.0.0

# Utilities
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
pillow>=9.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
